ability to display rule name in syntax diagram (AW-158)
[git-p4: depot-paths = "//depot/code/antlrworks/main/": change = 4960]

+++ src/org/antlr/works/visualization/graphics/primitive/GLiteral.java
@@ -98,7 +98,7 @@ public class GLiteral {
             do {
                 Float v = values.get(c());
                 if(v != null) {
-                    _value += v.floatValue();
+                    _value += v;
                 } else {
                     if(c().equals(OP_SUB))
                         _value -= evaluate_sub();
